scripts/requirements.txt
# Python dependencies for asset info generator
Pillow>=10.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
